requirements.txt
Flask
Werkzeug
SQLAlchemy
Flask-SQLAlchemy
python-dotenv
psycopg2-binary
azure-ai-inference
playwright
pydantic
beautifulsoup4
requests
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
